requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
SQLAlchemy==1.4.23
python-dotenv==1.0.1
numpy==1.24.3
scikit-learn==1.3.0
flask-login==0.6.2
flask-wtf==1.1.1
email-validator==2.0.0
gunicorn==21.2.0
